+++ newdocs/src/changelog.md
@@ -7,6 +7,8 @@
 
 - [Issue 1650](https://github.com/jackdewinter/pymarkdown/issues/1650)
     - Rules for Md049 (consistent emphasis) and Md050 (consistent strong emphasis)
+- [Issue 1652](https://github.com/jackdewinter/pymarkdown/issues/1652)
+    - Rule for Md051 (valid local link fragments)
 
 <!-- pyml disable-next-line no-duplicate-heading-->
 ### Fixed